refactor: 테스트 코드 패키지 정리 및 잘못된 이너 클래스 변경
