+++ src/test/java/domain/janggi/ScoreTest.java
@@ -1,9 +1,8 @@
-package domain;
+package domain.janggi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.janggi.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
