+++ src/test/java/domain/janggi/TurnTest.java
@@ -1,9 +1,7 @@
-package domain;
+package domain.janggi;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.janggi.Team;
-import domain.janggi.Turn;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
